move waypoint render method
